doit: Split out cmd_read
